requirements.txt
fastapi == 0.110.0
pydantic == 2.6.4
uvicorn == 0.29.0
sqlalchemy == 2.0.2
alembic == 1.13
psycopg2-binary == 2.9.9
psycopg2 == 2.9.9
asyncpg == 0.29.0
python-dotenv == 0.21.0
python-multipart == 0.0.9
pillow == 10.2.0
passlib == 1.7.4
python-jose == 3.3.0
uuid == 1.30
redis == 5.0.3
aioredis == 2.0.1
fastapi-cache2 == 0.2.0
celery == 5.4.0
flower == 2.0.0